extracting duplicate message helper usage to a module
